requirements.txt
requests==2.31.0
tweepy==4.14.0
praw==7.7.1
pandas==2.1.4
python-dotenv==1.0.0
click==8.1.7
textblob==0.17.1
beautifulsoup4==4.12.2
schedule==1.2.0
flask==2.3.3
flask-cors==4.0.0 